Support-1.5: Fix failing behats
